feat: migrate all monetary float fields to Decimal (Numeric 18,6)
- Update 9 schema files (~80+ fields) from float to Decimal
- Add Numeric(18,6) sa_columns for all SQLModel table fields
- Create Alembic migration for PostgreSQL column type changes
- Add DecimalSafeJSONResponse for backward-compatible JSON serialization
- Register Decimal->float in FastAPI ENCODERS_BY_TYPE
- Fix test assertions for Decimal/float pytest.approx compatibility
- plan_service.py simulation engine deferred to separate refactor

Closes #9

+++ apps/backend/app/schema/options_models.py
@@ -1,22 +1,23 @@
 from __future__ import annotations
 
+from decimal import Decimal
 from pydantic import BaseModel
 
 
 class OptionsRecord(BaseModel):
     year: int
-    amount: float
+    amount: Decimal
 
 
 class OptionsProjectionParams(BaseModel):
-    growth_rate: float
+    growth_rate: Decimal
     cutoff_year: int
     final_year: int
 
 
 class OptionsProjectionPoint(BaseModel):
     year: int
-    amount: float
+    amount: Decimal
     type: str  # "historical" or "projected"
